+++ backend/core/utils.py
@@ -530,7 +530,7 @@ def write_prism_file(date, payments, tomorrow, new_account_alias=False):
     output_dir = settings.PRISM_OUTPUT_DIR
 
     if new_account_alias:
-        filename_str = "_new"
+        filename_str = "_1"
     else:
         filename_str = ""
     # The microseconds are included to avoid accidentally overwriting
@@ -644,7 +644,7 @@ def export_prism_payments_for_date(date=None):
     return filename
 
 
-def generate_expected_payments_report_list():
+def generate_expected_payments_report_list(new_account_alias=False):
     """Generate a payments report of granted AND expected payments."""
     current_year = timezone.now().year
     two_years_ago = current_year - 2
@@ -667,11 +667,13 @@ def generate_expected_payments_report_list():
             "payment_schedule__activity__details",
         )
     )
-    payments_report_list = generate_payments_report_list(payments)
+    payments_report_list = generate_payments_report_list(
+        payments, new_account_alias
+    )
     return payments_report_list
 
 
-def generate_payments_report_list(payments):
+def generate_payments_report_list(payments, new_account_alias=False):
     """Generate a payments report list of payment dicts from payments."""
     payments_report_list = []
     for payment in payments:
@@ -731,6 +733,14 @@ def generate_payments_report_list(payments):
         mother = case.related_persons.filter(relation_type="mor").first()
         father = case.related_persons.filter(relation_type="far").first()
 
+        # TODO: the "_new" report should be the future default.
+        if new_account_alias:
+            account_string = payment.account_string_new
+            account_alias = payment.account_alias_new
+        else:
+            account_string = payment.account_string
+            account_alias = payment.account_alias
+
         payment_dict = {
             # payment specific.
             "id": payment.pk,
@@ -738,8 +748,8 @@ def generate_payments_report_list(payments):
             "paid_amount": payment.paid_amount,
             "date": payment.date,
             "paid_date": payment.paid_date,
-            "account_string": payment.account_string,
-            "account_alias": payment.account_alias,
+            "account_string": account_string,
+            "account_alias": account_alias,
             # payment_schedule specific.
             "payment_schedule__payment_id": payment_schedule.payment_id,
             "payment_schedule__"
@@ -787,6 +797,17 @@ def generate_payments_report_list(payments):
             "mother_cpr": mother.cpr_number if mother else None,
             "father_cpr": father.cpr_number if father else None,
         }
+
+        # TODO: Also add the activity_category to the new file.
+        category = activity.activity_category
+        category_name = activity.activity_category.name if category else None
+        category_id = (
+            activity.activity_category.category_id if category else None
+        )
+        if new_account_alias:
+            payment_dict["activity_category__category_id"] = category_id
+            payment_dict["activity_category__name"] = category_name
+
         payments_report_list.append(payment_dict)
 
     return payments_report_list
@@ -881,9 +902,9 @@ def validate_cvr(cvr):
     return bool(match)
 
 
-def parse_account_alias_mapping_data_from_csv(path):
+def parse_account_alias_mapping_data_from_csv_string(string):
     """
-    Parse account alias mapping data from a .csv file.
+    Parse account alias mapping data from a .csv StringIO.
 
     Returns a list of (main_account_number, activity_number, alias) tuples
     for example:
@@ -892,13 +913,14 @@ def parse_account_alias_mapping_data_from_csv(path):
         (528211011, 015038, BOS0000112)
     ]
     """
-    with open(path) as csvfile:
-        reader = csv.reader(csvfile)
-        rows = [row for row in reader]
+    reader = csv.reader(string)
+    rows = [row for row in reader]
 
     account_alias_mapping_data = []
 
     for row in rows:
+        if not len(row) > 1:
+            continue
         alias = row[0]
         account_string = row[1]
         if not alias or not alias.startswith("BOS"):
@@ -913,3 +935,12 @@ def parse_account_alias_mapping_data_from_csv(path):
         )
 
     return account_alias_mapping_data
+
+
+def parse_account_alias_mapping_data_from_csv_path(path):
+    """Helper-function for parsing account alias mappings from a path."""
+    with open(path) as csvfile:
+        account_alias_data = parse_account_alias_mapping_data_from_csv_string(
+            csvfile
+        )
+    return account_alias_data