+++ NEWS.rst
@@ -1,3 +1,18 @@
+Version 3.4.0, 2021-02-11
+-------------------------
+
+New in this version:
+
+Features
+^^^^^^^^
+
+- Add parallel account_string using the new activity category models.
+- Add parallel account_alias using the new account alias mapping models.
+- Update payment list when a payment is changed (for example paid).
+- Publish database documentation on build.
+- Update third party dependencies.
+
+
 Version 3.3.1, 2021-01-25
 -------------------------